// src/models/enums.ts
//Defining the color of the card
export enum CardColor {
  Red = "red",
  Green = "green",
  Blue = "blue",
  Yellow = "yellow",
  Wild = "wild",
}

//Defining the Characteristics of the card
export enum CardValue {
  Zero = "0",
  One = "1",
  Two = "2",
  Three = "3",
  Four = "4",
  Five = "5",
  Six = "6",
  Seven = "7",
  Eight = "8",
  Nine = "9",
  DrawTwo = "+2",
  WildCard = "wild",
  WildDrawFour = "+4",
  Reverse = "reverse",
  Block = "block",
  Wild = "wild",
}